chore: fix unit tests, reduce circular enum imports
