+++ src/Components/ServiceScene/Breakdowns/FieldsBreakdownScene.test.ts
@@ -47,7 +47,7 @@ describe('buildFieldsQueryString', () => {
 
     const result = buildFieldsQueryString('caller', filterVariable, detectedFieldsFrame);
     expect(result).toEqual(
-      `sum by (caller) (count_over_time({\${filters}}  \${levels} \${metadata} \${patterns} \${lineFilters} | logfmt | caller!="" \${fields} [$__auto]))`
+      `sum by (caller) (count_over_time({\${filters}}  \${levels} \${metadata} \${patterns} \${lineFilters} | logfmt | caller!="" \${fields} \${lineFormat} [$__auto]))`
     );
   });
   test('should build json-parser query', () => {
@@ -87,7 +87,7 @@ describe('buildFieldsQueryString', () => {
 
     const result = buildFieldsQueryString('caller', filterVariable, detectedFieldsFrame);
     expect(result).toEqual(
-      `sum by (caller) (count_over_time({\${filters}}  \${levels} \${metadata} \${patterns} \${lineFilters} | json | drop __error__, __error_details__ | caller!="" \${fields} [$__auto]))`
+      `sum by (caller) (count_over_time({\${filters}}  \${levels} \${metadata} \${patterns} \${lineFilters} | json \${jsonFields} | drop __error__, __error_details__ | caller!="" \${fields} \${lineFormat} [$__auto]))`
     );
   });
   test('should build mixed-parser query', () => {
@@ -126,7 +126,7 @@ describe('buildFieldsQueryString', () => {
 
     const result = buildFieldsQueryString('caller', filterVariable, detectedFieldsFrame);
     expect(result).toEqual(
-      `sum by (caller) (count_over_time({\${filters}}  \${levels} \${metadata} \${patterns} \${lineFilters} | json | logfmt | drop __error__, __error_details__ | caller!="" \${fields} [$__auto]))`
+      `sum by (caller) (count_over_time({\${filters}}  \${levels} \${metadata} \${patterns} \${lineFilters} | json \${jsonFields} | logfmt | drop __error__, __error_details__ | caller!="" \${fields} \${lineFormat} [$__auto]))`
     );
   });
   test('should build metadata query', () => {
@@ -165,7 +165,7 @@ describe('buildFieldsQueryString', () => {
 
     const result = buildFieldsQueryString('caller', metadataVariable, detectedFieldsFrame);
     expect(result).toEqual(
-      `sum by (caller) (count_over_time({\${filters}} | caller!="" \${levels} \${metadata} \${patterns} \${lineFilters}  \${fields} [$__auto]))`
+      `sum by (caller) (count_over_time({\${filters}} | caller!="" \${levels} \${metadata} \${patterns} \${lineFilters}  \${fields} \${lineFormat} [$__auto]))`
     );
   });
 });
