+++ src/__mocks__/@gtk-grafana/react-json-tree.tsx
@@ -1,6 +1,5 @@
-import { JSONTree } from '@gtk-grafana/react-json-tree/dist';
 import React from 'react';
 
 module.exports = {
-  // JSONTree: <></>
+  JSONTree: <></>,
 };
