+++ src/services/routing.test.ts
@@ -1,5 +1,6 @@
-import { buildServicesUrl, PageSlugs, ROUTES } from './routing';
+import { buildServicesUrl, ROUTES } from './routing';
 import { buildDrilldownPageUrl } from './navigate';
+import { PageSlugs } from './enums';
 
 describe('buildBreakdownUrl', () => {
   const OLD_LOCATION = window.location;
